moved database folder due to tsc conflicts
