+++ build.ts
@@ -4,6 +4,5 @@ const outDir = config.compilerOptions.outDir;
 
 s.rm('-rf', outDir);
 s.mkdir(outDir);
-s.cp('.env', `${outDir}/.env`);
 s.mkdir('-p', `${outDir}/common/swagger`);
 s.cp('server/common/api.yml', `${outDir}/common/api.yml`);
