+++ package.json
@@ -10,7 +10,7 @@
     "dev": "nodemon server/index.ts | pino-pretty",
     "dev:debug": "nodemon --exec 'node -r ts-node/register --inspect-brk' server/index.ts | pino-pretty",
     "compile": "ts-node build.ts && tsc",
-    "create:schema": "ts-node -r dotenv/config database/scripts/create_db.ts",
+    "create:schema": "ts-node -r dotenv/config server/database/scripts/create_db.ts",
     "start": "node dist/index.js",
     "test": "mocha test/**/*.ts --exit",
     "test:debug": "mocha --inspect-brk test/**/*.ts --exit",
