+++ server/database/models/index.ts
@@ -1,4 +1,4 @@
 export {TCDocument} from './tcDocument.model';
 export {TCUserAgreement} from './tcUserAgreement.model';
-export {TCDocumentApp} from 'tcDocumentApp.model';
+export {TCDocumentApp} from './tcDocumentApp.model';
 export {TCApp} from './tcApp.model';
