+++ server/database/models/tcUserAgreement.model.ts
