+++ server/database/repos/TCAppRepository.ts
@@ -1,16 +1,16 @@
 import {IDatabase, IMain} from 'pg-promise';
 import {IResult} from 'pg-promise/typescript/pg-subset';
-import {TCUserAgreement} from '../models';
-import {userAgreements as sql} from '../sql';
-import {TCColumnSets} from "../models/tcColumnSet.model";
+import {TCDocumentApp} from '../models';
+import {apps as sql} from '../sql';
+import {TCColumnSets} from '../models/tcColumnSet.model';
 
-export class TCUserAgreementRepository {
+export class TCAppRepository {
 
     constructor(private db: IDatabase<any>, private pgp: IMain) {
         this.createColumnSets();
     }
 
-    private static table = 'TCUserAgreement';
+    private static table = 'TCApp';
 
     // ColumnSet objects static namespace:
     private static cs: TCColumnSets;
@@ -20,38 +20,38 @@ export class TCUserAgreementRepository {
         return this.db.none(sql.create);
     }
 
-    // Adds a new user, and returns the new object;
-    async add(name: string): Promise<TCUserAgreement> {
-        return this.db.one(sql.add, name);
+    // Adds a document app
+    async add(app: string): Promise<TCDocumentApp> {
+        return this.db.one(sql.add, app);
     }
 
     // Tries to delete a user by id, and returns the number of records deleted;
     async remove(id: number): Promise<number> {
-        return this.db.result(`DELETE FROM ${TCUserAgreementRepository.table} WHERE id = $1`, +id, (r: IResult) => r.rowCount);
+        return this.db.result(`DELETE FROM ${TCAppRepository.table} WHERE id = $1`, +id, (r: IResult) => r.rowCount);
     }
 
     // Returns all user records;
-    async all(): Promise<TCUserAgreement[]> {
-        return this.db.any(`SELECT * FROM ${TCUserAgreementRepository.table}`);
+    async all(): Promise<TCDocumentApp[]> {
+        return this.db.any(`SELECT * FROM ${TCAppRepository.table}`);
     }
 
     // Returns the total number of users;
     async total(): Promise<number> {
-        return this.db.one(`SELECT count(*) FROM ${TCUserAgreementRepository.table}`, [], (a: { count: string }) => +a.count);
+        return this.db.one(`SELECT count(*) FROM ${TCAppRepository.table}`, [], (a: { count: string }) => +a.count);
     }
 
     // example of setting up ColumnSet objects:
     private createColumnSets(): void {
         // create all ColumnSet objects only once:
-        if (!TCUserAgreementRepository.cs) {
+        if (!TCAppRepository.cs) {
             const helpers = this.pgp.helpers, cs: TCColumnSets = {};
 
-            const table = new helpers.TableName({table: TCUserAgreementRepository.table, schema: 'public'});
+            const table = new helpers.TableName({table: TCAppRepository.table, schema: 'public'});
 
             cs.insert = new helpers.ColumnSet(['name'], {table});
             cs.update = cs.insert.extend(['?id']);
 
-            TCUserAgreementRepository.cs = cs;
+            TCAppRepository.cs = cs;
         }
     }
