+++ server/database/repos/TCUserAgreementRepository.ts
@@ -1,16 +1,16 @@
 import {IDatabase, IMain} from 'pg-promise';
 import {IResult} from 'pg-promise/typescript/pg-subset';
-import {TCDocumentApp} from '../models';
-import {documentApps as sql} from '../sql';
-import {TCColumnSets} from "../models/tcColumnSet.model";
+import {TCUserAgreement} from '../models';
+import {userAgreements as sql} from '../sql';
+import {TCColumnSets} from '../models/tcColumnSet.model';
 
-export class TCDocumentAppRepository {
+export class TCUserAgreementRepository {
 
     constructor(private db: IDatabase<any>, private pgp: IMain) {
         this.createColumnSets();
     }
 
-    private static table = 'TCDocumentApp';
+    private static table = 'TCUserAgreement';
 
     // ColumnSet objects static namespace:
     private static cs: TCColumnSets;
@@ -20,38 +20,38 @@ export class TCDocumentAppRepository {
         return this.db.none(sql.create);
     }
 
-    // Adds a document app
-    async add(documentId: number, app: string): Promise<TCDocumentApp> {
+    // Adds a new user, and returns the new object;
+    async add(name: string): Promise<TCUserAgreement> {
         return this.db.one(sql.add, name);
     }
 
     // Tries to delete a user by id, and returns the number of records deleted;
     async remove(id: number): Promise<number> {
-        return this.db.result(`DELETE FROM ${TCDocumentAppRepository.table} WHERE id = $1`, +id, (r: IResult) => r.rowCount);
+        return this.db.result(`DELETE FROM ${TCUserAgreementRepository.table} WHERE id = $1`, +id, (r: IResult) => r.rowCount);
     }
 
     // Returns all user records;
-    async all(): Promise<TCDocumentApp[]> {
-        return this.db.any(`SELECT * FROM ${TCDocumentAppRepository.table}`);
+    async all(): Promise<TCUserAgreement[]> {
+        return this.db.any(`SELECT * FROM ${TCUserAgreementRepository.table}`);
     }
 
     // Returns the total number of users;
     async total(): Promise<number> {
-        return this.db.one(`SELECT count(*) FROM ${TCDocumentAppRepository.table}`, [], (a: { count: string }) => +a.count);
+        return this.db.one(`SELECT count(*) FROM ${TCUserAgreementRepository.table}`, [], (a: { count: string }) => +a.count);
     }
 
     // example of setting up ColumnSet objects:
     private createColumnSets(): void {
         // create all ColumnSet objects only once:
-        if (!TCDocumentAppRepository.cs) {
+        if (!TCUserAgreementRepository.cs) {
             const helpers = this.pgp.helpers, cs: TCColumnSets = {};
 
-            const table = new helpers.TableName({table: TCDocumentAppRepository.table, schema: 'public'});
+            const table = new helpers.TableName({table: TCUserAgreementRepository.table, schema: 'public'});
 
             cs.insert = new helpers.ColumnSet(['name'], {table});
             cs.update = cs.insert.extend(['?id']);
 
-            TCDocumentAppRepository.cs = cs;
+            TCUserAgreementRepository.cs = cs;
         }
     }
