+++ tsconfig.json
@@ -9,6 +9,6 @@
     "outDir": "dist",
     "typeRoots": ["node_modules/@types"]
   },
-  "include": ["typings.d.ts", "server/**/*.ts", "database/**/*.ts"],
+  "include": ["typings.d.ts", "server/**/*.ts"],
   "exclude": ["node_modules"]
 }
